Write a script to schedule radio music.
Why?  Because listening to complete albums is nice (often tracks build
off each other), but variety is also nice.  Furthermore, having an
actual in-order playlist, rather than random shuffle, means it could
be displayed on the website at some point.

How It Works
------------

Music is scheduled in four-hour slots, consisting of:

1. A transition track ("you're listening to lainchan radio", etc)
2. A full album
3. Enough tracks to make up the difference
4. A full album

Scheduling of the random tracks is greedy, so this may not be totally
right.  But the radio seems to have enough variety that it gets within
a couple of seconds every time I've tried it.

Thoughts
--------

A possible future change is imposing a maximum length of an album.
For instance, the "Bootleg Lain" album is over an hour long, and the
tracks are all pretty much unrelated other than genre.  Alternatively,
maybe that music needs correctly categorising.

Another possible change is increasing the slot size.  Initially I
tried two hours, but only one or two random tracks managed to fit
between the albums.  Four hours is much better, but the generated
programme is still dominated by the albums.