Pass "now" parameter to icecast_metrics_list.
This change is unnecessary, but greatly clarifies what is happening.
The old code worked due two aspects of Python's scoping rules:

1. Globals in Python can be read in a function.  It's only if you want
   to write that you need `global`.
2. Python has no block scope.

(2) means that `now` is in global scope, despite being buried inside
two levels of indentation.  (1) means that `icecast_metrics_list` can
refer to it even though it's not a parameter.

Python is stupid and confusing.