Start mpd playing if it's not already
MPD should always be playing, but today it stopped at 7AM and didn't
start again.

An alternative approach would be to introduce a separate systemd unit
to monitor MPD and give it a kick if it's not in the right state, but
since not playing triggers an error from the backend, it can be fixed
here too.